# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "E-commerce API",
        "description": "API for managing products, categories, users, and cart items in an e-commerce application."
    },
    "servers": [
        {
            "url": "http://localhost:3200"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": ["Products"],
                "summary": "Get all products",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "A list of products"
                    }
                }
            },
            "post": {
                "tags": ["Products"],
                "summary": "Create a new product",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "price": { "type": "number" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    }
                }
            }
        },
        "/api/users/signin": {
            "post": {
                "tags": ["Users"],
                "summary": "User Sign In",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": { "type": "string" },
                                    "password": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": { "description": "Successful sign in" },
                    "401": { "description": "Incorrect credentials" }
                }
            }
        },
        "/api/users/signup": {
            "post": {
                "tags": ["Users"],
                "summary": "User Sign Up",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "email": { "type": "string" },
                                    "password": { "type": "string" },
                                    "type": { "type": "string" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": { "description": "User created successfully" },
                    "400": { "description": "Invalid input" }
                }
            }
        }
    }
}